# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseStatusView": {
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_reference": {
            "title": "Case Reference",
            "type": "string"
          },
          "final_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewDecisionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "possible_next_statuses": {
            "items": {
              "$ref": "#/components/schemas/CaseStatus"
            },
            "title": "Possible Next Statuses",
            "type": "array"
          },
          "review_queue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewQueue"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "case_reference",
          "status",
          "review_queue",
          "possible_next_statuses",
          "final_decision",
          "updated_at"
        ],
        "title": "CaseStatusView",
        "type": "object"
      },
      "CaseView": {
        "properties": {
          "assigned_reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Reviewer Id"
          },
          "case_reference": {
            "title": "Case Reference",
            "type": "string"
          },
          "clinical_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Summary"
          },
          "close_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CloseReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "conservative_treatment_weeks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conservative Treatment Weeks"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "diagnosis_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis Code"
          },
          "diagnosis_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis Description"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentView"
            },
            "title": "Documents",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "patient": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientView"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyView"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderView"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_service": {
            "$ref": "#/components/schemas/RequestedServiceView"
          },
          "review_queue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewQueue"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Requested At"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "urgency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Urgency"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "case_reference",
          "status",
          "urgency",
          "provider",
          "patient",
          "policy",
          "requested_service",
          "diagnosis_code",
          "diagnosis_description",
          "conservative_treatment_weeks",
          "clinical_summary",
          "documents",
          "review_queue",
          "assigned_reviewer_id",
          "review_requested_at",
          "close_reason",
          "created_at",
          "updated_at",
          "closed_at",
          "version"
        ],
        "title": "CaseView",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HumanDecisionCommand": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/HumanDecisionType"
          },
          "rationale": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Rationale",
            "type": "string"
          },
          "recommendation_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Recommendation Id",
            "type": "string"
          }
        },
        "required": [
          "recommendation_id",
          "decision",
          "rationale"
        ],
        "title": "HumanDecisionCommand",
        "type": "object"
      },
...
      "ReviewPacketView": {
        "properties": {
          "audit_history": {
            "items": {
              "$ref": "#/components/schemas/AuditEventView"
            },
            "title": "Audit History",
            "type": "array"
          },
          "case": {
            "$ref": "#/components/schemas/CaseView"
          },
          "current_recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/ReviewDecisionView"
            },
            "title": "Decisions",
            "type": "array"
          },
          "recommendation_history": {
            "items": {
              "$ref": "#/components/schemas/RecommendationView"
            },
            "title": "Recommendation History",
            "type": "array"
          }
        },
        "required": [
          "case",
          "current_recommendation",
          "recommendation_history",
          "decisions",
          "audit_history"
        ],
        "title": "ReviewPacketView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nBackend for provider pre-authorisation requests: case intake, validation, rule evaluation, advisory\nrecommendations, human review, and an immutable audit trail.\n\n**The system never issues final authorisations or denials.** Only an authorised human reviewer can move a case\nto `APPROVED` or `DENIED`.\n\n**Authentication assumption:** requests arrive through a gateway that authenticates the caller and sets\n`X-Actor-Type`, `X-Actor-Id`, and (for reviewers) `X-Actor-Roles`. The service must not be exposed directly.\n\nEvery response carries `X-Request-ID`. Errors use a uniform envelope: `{\"error\": {\"code\", \"message\", \"details\",\n\"request_id\", \"case_id\"}}`.\n",
    "title": "Pre-Authorisation Case Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases/{case_id}": {
      "get": {
        "description": "Returns all collected information, documents, and current status.\n\n**Authorisation:** Any authenticated actor.\n\n**State transitions:** None.",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Retrieve a case",
        "tags": [
          "Provider interaction"
        ]
      }
    },
...
    "/api/v1/review/cases/{case_id}": {
      "get": {
        "description": "Everything a reviewer needs: case details, provider request, collected information, documents, current and historical recommendations with rule results and evidence, prior decisions, and full audit history.\n\n**Authorisation:** Actor type `HUMAN_REVIEWER`.\n\n**State transitions:** None.",
        "operationId": "review_packet_api_v1_review_cases__case_id__get",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPacketView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HUMAN_REVIEWER_REQUIRED"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Retrieve the review packet",
        "tags": [
          "Human review"
        ]
      }
    },
    "/api/v1/review/cases/{case_id}/assignment": {
      "post": {
        "description": "Self-assignment. A decision can only be recorded by the assigned reviewer. Reassignment is allowed and audited with the previous assignee.\n\n**Authorisation:** Actor type `HUMAN_REVIEWER` with role `CLINICAL_REVIEWER` (case `PENDING_HUMAN_REVIEW`) or `MEDICAL_DIRECTOR` (case `ESCALATED`).\n\n**State transitions:** None (records `REVIEWER_ASSIGNED`).",
        "operationId": "assign_api_v1_review_cases__case_id__assignment_post",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStatusView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HUMAN_REVIEWER_REQUIRED / INSUFFICIENT_REVIEWER_ROLE"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_UNDER_REVIEW / CONCURRENT_MODIFICATION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Assign the case to yourself",
        "tags": [
          "Human review"
        ]
      }
    },
    "/api/v1/review/cases/{case_id}/decision": {
      "post": {
        "description": "Records the reviewer's decision against the current recommendation (`recommendation_id` must be the latest; guards against deciding on stale information). A rationale is mandatory. The system recommendation is never modified; departures are flagged `is_override` and audited as `RECOMMENDATION_OVERRIDDEN`.\n\n**Authorisation:** Actor type `HUMAN_REVIEWER` with role `CLINICAL_REVIEWER` (case `PENDING_HUMAN_REVIEW`) or `MEDICAL_DIRECTOR` (case `ESCALATED`). Must be the assigned reviewer.\n\n**State transitions:** `PENDING_HUMAN_REVIEW` \u2192 `APPROVED` / `DENIED` / `PENDING_INFORMATION` / `ESCALATED`; `ESCALATED` \u2192 `APPROVED` / `DENIED` / `PENDING_INFORMATION`.",
        "operationId": "record_decision_api_v1_review_cases__case_id__decision_post",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanDecisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HUMAN_REVIEWER_REQUIRED / INSUFFICIENT_REVIEWER_ROLE / REVIEWER_NOT_ASSIGNED"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_UNDER_REVIEW / STALE_RECOMMENDATION / INVALID_STATE_TRANSITION / CONCURRENT_MODIFICATION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Record a human decision",
        "tags": [
          "Human review"
        ]
      }
    },
    "/api/v1/review/queues/{queue}": {
      "get": {
        "description": "Cases awaiting review in the queue; expedited first, then oldest first.\n\n**Authorisation:** Actor type `HUMAN_REVIEWER`.\n\n**State transitions:** None.",
        "operationId": "review_queue_api_v1_review_queues__queue__get",
        "parameters": [
          {
            "in": "path",
            "name": "queue",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReviewQueue"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItemView"
                  },
                  "title": "Response Review Queue Api V1 Review Queues  Queue  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HUMAN_REVIEWER_REQUIRED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "List a review queue",
        "tags": [
          "Human review"
        ]
      }
    },
...
  }
}